(HP-1031) Moved common.ts from client/__tests__/
Moved the file so it does not need to have a test. Removed the test too. The test was run in every *.test.ts which imported the common.ts.
