+++ package.json
@@ -33,6 +33,7 @@
     "http-status-typed": "^1.0.0",
     "jest-fetch-mock": "3.0.3",
     "jwt-decode": "^3.0.0",
+    "mocked-env": "^1.3.2",
     "oidc-client": "1.11.5",
     "react": "^16.11.0",
     "react-dom": "16.13.1",
@@ -72,7 +73,6 @@
     "@testing-library/react": "^11.0.4",
     "jest-environment-jsdom-sixteen": "^1.0.3",
     "jest-sonar-reporter": "^2.0.0",
-    "mocked-env": "^1.3.2",
     "prettier": "^1.18.2",
     "ts-node": "^8.8.2"
   },
