+++ src/tests/client.test.helper.ts
@@ -1,8 +1,13 @@
 import { FetchMock } from 'jest-fetch-mock';
 import mockedEnv, { RestoreFn } from 'mocked-env';
-import { Client, FetchApiTokenOptions, getClientConfig, getTokenUri } from '..';
-import { MockMutator, promiseDefaultTimeout } from '../__mocks__/index';
-import { AnyFunction, AnyObject } from '../../common';
+import {
+  Client,
+  FetchApiTokenOptions,
+  getClientConfig,
+  getTokenUri
+} from '../client';
+import { MockMutator, promiseDefaultTimeout } from '../client/__mocks__/index';
+import { AnyFunction, AnyObject } from '../common';
 
 type NodeJS = AnyObject;
 
@@ -83,9 +88,3 @@ export const createApiTokenFetchPayload = (
 
 export const setEnv = (overrides: Partial<NodeJS.ProcessEnv>): RestoreFn =>
   mockedEnv(overrides, { restore: true });
-
-describe('createApiTokenFetchPayload', () => {
-  it('returns an object', () => {
-    expect(createApiTokenFetchPayload()).toBeDefined();
-  });
-});
